rafactor: move base part into parts module
