+++ src/ansys/health/heart/writer/base_writer.py
@@ -31,9 +31,9 @@ import pandas as pd
 
 from ansys.dyna.core.keywords import keywords
 from ansys.health.heart import LOG as LOGGER
-from ansys.health.heart.anatomical_parts import _Chamber
+from ansys.health.heart.anatomical_parts import PartType, _Chamber
 from ansys.health.heart.models import BiVentricle, FourChamber, FullHeart, HeartModel, LeftVentricle
-from ansys.health.heart.objects import PartType, SurfaceMesh
+from ansys.health.heart.objects import SurfaceMesh
 import ansys.health.heart.settings.settings as sett
 from ansys.health.heart.settings.settings import SimulationSettings
 from ansys.health.heart.writer import custom_keywords as custom_keywords
