+++ tests/heart/writer/test_dynawriter.py
@@ -28,11 +28,10 @@ import pytest
 import pyvista as pv
 import pyvista.examples as examples
 
+from ansys.health.heart.anatomical_parts import Part, PartType
 from ansys.health.heart.models import FullHeart
 from ansys.health.heart.objects import (
     Mesh,
-    Part,
-    PartType,
     Point,
 )
 from ansys.health.heart.pre.conduction_path import ConductionPath, ConductionPathType
